Add error handler that prints JSON to stdout
As virtomate's output is JSON, it makes sense to return errors as JSON, too. So
far, the error message is limited to the type of error (exception name) and a
message. If necessary, more information can be added in the future.

The errors are printed to stdout and not stderr. By printing all JSON to stdout,
virtomate can use stderr to print the help, log messages, or debug output.
Users can determine whether to expect a normal message or an error by
inspecting virtomate's exit code.