Store test volumes in default storage pool
The idea behind the separate storage pool was to prevent naming clashes with
existing volumes when running tests against libvirt on the developer's machine.
But that turned out to be too much hassle. In addition to creating and
maintaining the separate storage pool, developers would have to reconfigure
AppArmor on Debian and Ubuntu or the volumes were inaccessible.

In the future, developers should create a dedicated VM for testing virtomate.
This also eliminates the risk of accidentally deleting the developer's data.