Retain volume format when cloning EFI firmware
Some systems, notably Fedora 40, include EFI firmware on qcow2 volumes. When
copying such a volume, libvirt needs to be told in what format the volume is,
otherwise it will assume that it is a raw volume and use the wrong driver to
access the volume. The previous code to copy firmware did not take this into
account. Consequently, cloned machines would not boot.

Unfortunately, support for EFI firmware on qcow2 volumes is scarce. Neither
Debian 12 nor Ubuntu 24.04 have it. Packer only supports firmware on raw
volumes, too. So there is no realistic way to produce a VM with firmware on
qcow2 for testing purposes. That's why there is no test.