Guard against crashes after enumerating objects
If a libvirt object (domain, pool, volume, ...) gets removed after obtaining a
handle, accessing its properties causes libvirt to throw exceptions, which in
turn causes Virtomate operations to fail. That is desired while working with
the respective object. For example, removing the domain that is currently being
cloned is irrecoverable. Hence, the cloning process must be aborted. However,
it is disruptive and unnecessary when the removal of an *unrelated* object
causes the current operation to fail. For example, listing volumes should
succeed despite a volume having been removed.

This commit guards against crashes in the most obvious places (listing domains,
pools, volumes) by logging the exception and continuing. Cloning was also
affected by this problem: We check that randomly generated MAC addresses and
UUIDs are unique by comparing them with all domains on a machine. Again,
exceptions will be logged and otherwise ignored.